[XLA] Avoid harmless signed integer overflow
A U16 of 46977 multiplied by a U16 of 53826, when evaluated in the evaluator,
results in the operands of the multiply getting promoted to the C++ type "int"
which is signed. The result of the multiply will overflow a signed int and give
a negative result.

Instead, let's promote both operands to "unsigned int" which will not suffer
from any overflow issues.

PiperOrigin-RevId: 325291178
Change-Id: Ib20b9b56fe3803e7ec4a927abb6b862683523068