Make ScopedAllocatorOptimizer compatible with Const input.
In a previous change, we aborted ScopedAllocatorOptimizer when one of the inputs
is a Const op.

This change instead enables Const op to work with ScopedAllocatorOptimizer by
introducing an Identity op after Const.  Thus we change:
Const -> CollectiveReduce
to
Const -> Identity -> CollectiveReduce

The Identity becomes the real input to CollectiveReduce, and it will use the
pre-allocated buffer slice for its output tensor when it invokes `set_output`
by this logic:
https://github.com/tensorflow/tensorflow/blob/6b65afa4209a8743d819693ab6c50b5df4db8af9/tensorflow/core/framework/op_kernel.cc#L903.

This is similar to the approach in cl/259138773.

PiperOrigin-RevId: 325541732
Change-Id: I6487685089520b73387197a31fef5780217a3a4b