Get namedtuple _make method from instance instead of class.
It is possible that v0 is a _TupleWrapper object wrapping a namedtuple instead of a namedtuple itself. The class _TupleWrapper does not have the class method _make, but it should have the instance method _make which calls the namedtuple class method.

Instances of a namedtuple also have the instance method _make which calls the relevant class method. This will allow the code to work for both cases.

PiperOrigin-RevId: 325815594
Change-Id: I209f8ec5a8617f72183e4c12937b4429321e7b4f