[XLA] NFC: Refactor allocation success/failure logic in memory space assignment.
This CL is in preparation for repacking the allocations in case we run out of
alternate memory due to fragmentation. This CL introduces a Result enum to
capture success of an allocation or failure due to multiple different reasons.
Moved the logic around to perform the decision to Finalize or Uncommit (and
possibly retry) the allocations based on the Result. We can further use this
Result to repack the assigned offsets to help with fragmentation.

I also modified the function signatures that I touched to use non-const
references instead of non-const pointers due to go/totw/178.

PiperOrigin-RevId: 325881001
Change-Id: Icdbf1892d75208ec3eba8d2e084097106fb07dde