[tf.data service] Share dataset graphs via filesystem with RPC fallback.
If a `work_dir` is configured, the dispatcher will write datasets there. When a worker needs a dataset, the dispatcher will send the worker the filesystem path of the dataset. The worker will attempt to load the dataset from filesystem. If the file reading fails, the worker falls back to sending the dispatcher an RPC request for the full dataset graph.

To aid in the implementation, this CL switches worker_impl to use a DataServiceDispatcherClient instead of a raw DataService::Stub, avoiding grpc boilerplate in worker_impl.

PiperOrigin-RevId: 327332591
Change-Id: I8b674d3391c7b78cb17ffd431da6eaa0e24ee30a