Fold Identity op in TF to CoreRT dialect lowering pass.
It's OK to fold Identity op for now, since tf_to_corert pass only supports single device placement today. When we support device placer in lowering pass, we need to check if the identity op and it's previous op are placed on the same device to decide if it is foldable.

PiperOrigin-RevId: 327870203
Change-Id: If43177625fbd139fb5304a94bd54d35fa5e41417