Parallel device: Remove the DeviceID op usage for now
It may be good to have later, but for now having an op registered is an unnecessary API surface. There's not a terribly strong case for the op yet, and having it as a dependency but not a core op makes it hard to depend on the parallel device.

Keeps a "device_id" property on the ParallelDevice in Python, which will behave identically to the previous version

PiperOrigin-RevId: 329313676
Change-Id: Iaf024d0d762061ba90e3568199334b53432001ce