Fix tests with bad saved models
We were only testing that the model fails to load due to being invalid but we were not looking at the root cause for the failure. Turns out the status was an error only because the model could not be loaded because it was not in the right format. In fact, the test data was not even included as dependency of the target, so the test could not even find the file.

The empty file in `assets` is because source control needs a file to track a directory.

PiperOrigin-RevId: 329389050
Change-Id: Ie484e7b9dd7675a29f5c0d70115d1e3cafd3d5f8