Add a generic build file and debug_log for Cortex M4 targets.
This generic target and debug_log implementation can be used to generate a microlite lib that teams can use to link against when including TFLM in their firmware.

Simply pass in the path to the desired toolchain with |TOOLCHAIN_PATH| and the Makefile will build with that version of the ARM GCC toolchain.

PiperOrigin-RevId: 329770063
Change-Id: If0b3f605c6ffb310424a41d773ae97b4672bb0a0