Do not clear the shape when using checkpoint_initializer to create variables.
It seems unnecessary to clear the shape. If the shape is incompatible with the checkpoint, variable creation would throw an error.

Keeping the shape allows outer variable_creator_scope to use the shape in meaningful ways, e.g, sharded variable creator can easily partition variables if the shape is known.

PiperOrigin-RevId: 330966090
Change-Id: I92441d8c378c73e12c20fec0a3a3b070d175b7ad