[tf.data service] Garbage collect old and unused jobs.
This CL adds a worker heartbeat which does the following:
- Registers the worker with the dispatcher if the worker is not yet registered
- Reports to the dispatcher which tasks the worker is currently processing
- Learns from the dispatcher which tasks should be added and which should be deleted.

Learning about new tasks is important in case the dispatcher's original ProcessTask request to the worker failed due to network issues. The heartbeat provides a backup mechanism where the worker can still learn about what tasks it should process.

Deleting old tasks is important for job lifecycle management. When the dispatcher
detects that a job is old and unused, it will mark the job (and all of its tasks) as finished. The worker will learn about the task finishing during the next heartbeat.

PiperOrigin-RevId: 330990312
Change-Id: Ifc0dbb3465ff478de83bfb3a265d9a57072d852a