Do not symlink system protobuf headers but only required .proto files
Symlinking the system headers has proven to be problematic as newer
versions of protobuf add or remove headers which makes having a static
array of header files hard to impossible. Turns out the headers don't
need to be symlinked at all but only the .proto files used as inputs
need to be present.