Simplify (compile-time) selective registration
This change eliminates some redundancy and complexity in the existing
'selective registration' implementation, in preparation for replacing it.

If selective registration is not enabled (the usual case), this change has no
behavioral impact.

If selective registration is enabled, there is a potential behavioral change:
  - A kernel with implementation 'K' is enabled (via SHOULD_REGISTER_OP_KERNEL("K"))
  - Its corresponding op 'A' is not enabled (via SHOULD_REGISTER_OP("A"))
(this is expected to be fixed, in a subsequent change)

Currently, such kernels would not be registered at runtime (see the magic
string "_no_register"), but will likely be included in the binary (i.e. the
workaround is not helpful for binary size). However, this involves a use of
SHOULD_REGISTER_OP on a non-constant expression (the parameter to Name()),
perhaps leading to its implementation appearing in the final binary. With this
change, such kernels would be registered (perhaps causing an error to be logged).

PiperOrigin-RevId: 331859457
Change-Id: I3726e3f9f67aeec3169a0927fe358dc2637b2396