[TF:TRT] Fix get_tensorrt_rewriter_config.
Previously, the routine only passes max_batch_size to the TF1
TensorRTOptimizer. The change is to passes max_batch_size to the
TensorRTOptimizer regardless the Tensorflow version, so that the value is
available when use_implicit_batch is true.

Fix the document for this.

Modify an existing test to test the passing of max_batch_size for TF2.

PiperOrigin-RevId: 332288040
Change-Id: I7079633427cba23bac1362469f49c698b23ab505